src/requirements.txt
boto3==1.37.29
botocore==1.37.29
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
docker==7.1.0
et_xmlfile==2.0.0
greenlet==3.1.1
idna==3.10
jmespath==1.0.1
jwt==1.3.1
MarkupSafe==3.0.2
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-http-client==3.3.7
pytz==2025.2
requests==2.32.3
s3transfer==0.11.4
sendgrid==6.11.0
six==1.17.0
smart-open==7.1.0
SQLAlchemy==2.0.40
starkbank-ecdsa==2.2.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
xlrd==2.0.1
xmltodict==0.14.2
